refactor: nav 타입 파일 이름 변경
